requirements.txt
fastapi
uvicorn
python-dotenv
openai
langchain
langchain-openai
langchain-community
chromadb
